requirements.txt
comtypes==1.2.0
docx==0.2.4
docx2pdf==0.1.8
fpdf==1.7.2
pandas==2.0.3
pdf2docx==0.5.6
pdf2image==1.16.3
pdfdocument==4.0.0
Pillow==9.4.0
Pillow==10.0.0
pypdf==3.13.0
PyPDF2==3.0.1
python_docx==0.8.11
pywin32==306
reportlab==4.0.4
tabula==1.0.5
tabula_py==2.7.0
tk==0.1.0
